Refactor module for class.js in modules file
