+++ index.js
@@ -1,7 +1,7 @@
 import {
-    bookSection,
-    titleText,
-    authorText,
+    // bookSection,
+    // titleText,
+    // authorText,
     form,
     listbtn,
     addlistbtn,
@@ -12,43 +12,12 @@ import {
     date,
 } from "./modules/book.js";
 
-class Book {
-  constructor() {
-    this.bookLibrary = JSON.parse(localStorage.getItem("jsonLibrary")) || [];
-  }
+// showClass.js
+import Book from "./modules/class.js";
+// window.Function = () => {
+//   bookList.remove("${n.id}");
+// };
 
-  insertHtml() {
-    bookSection.innerHTML = "";
-    this.bookLibrary.forEach((n) => {
-      bookSection.innerHTML += `<div class='book'>
-        <div class='bookDetail'>
-          <p>"${n.title}"</p>
-          <p>by</p>
-          <p>${n.author}</p>
-        </div>    
-          <button class='remove' type='button' onclick='bookList.remove("${n.id}")'>Remove</button>
-      </div>`;
-    });
-  }
-
-  addBook() {
-    const bookObject = {};
-    bookObject.title = titleText.value;
-    bookObject.author = authorText.value;
-    bookObject.id = (titleText.value + authorText.value).replace(/\s/g, "");
-    this.bookLibrary.push(bookObject);
-    localStorage.setItem("jsonLibrary", JSON.stringify(this.bookLibrary));
-    this.insertHtml();
-  }
-
-  remove(idParameter) {
-    this.bookLibrary = this.bookLibrary.filter(
-      (book) => book.id !== idParameter
-    );
-    localStorage.setItem("jsonLibrary", JSON.stringify(this.bookLibrary));
-    this.insertHtml();
-  }
-}
 
 const bookList = new Book();
 
@@ -102,3 +71,7 @@ function clock() {
 }
 
 setInterval(clock, 1000);
+
+const removebtn = document.querySelectorAll(".remove");
+
+console.log(removebtn.length);
